fix: Fix tests

+++ src/main/java/org/isf/menu/service/MenuIoOperations.java
@@ -204,11 +204,11 @@ public class MenuIoOperations {
 	 */
 	public void deleteUser(User user) throws OHServiceException {
 		ensureUserNotDeleted(user.getUserName());
-		try {
-			repository.delete(user);
-		} catch (Exception ex) {
+		if (user.isDeleted()) {
 			user.setDeleted(true);
 			repository.save(user);
+		} else {
+			repository.delete(user);
 		}
 	}
